Conflicts:
	elza-ui/src/main/java/cz/tacr/elza/ui/view/FindingAidDetailView.java

+++ elza-core/src/main/java/cz/tacr/elza/controller/PartyManager.java
@@ -4,11 +4,13 @@ import cz.tacr.elza.domain.ParAbstractParty;
 import cz.tacr.elza.domain.ParAbstractPartyVals;
 import cz.tacr.elza.domain.ParPartySubtype;
 import cz.tacr.elza.domain.ParPartyType;
+import cz.tacr.elza.domain.ParPartyTypeExt;
 import cz.tacr.elza.domain.RegRecord;
 import cz.tacr.elza.repository.AbstractPartyRepository;
 import cz.tacr.elza.repository.PartySubtypeRepository;
 import cz.tacr.elza.repository.PartyTypeRepository;
 import cz.tacr.elza.repository.RegRecordRepository;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.Assert;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,7 +21,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Nullable;
 import javax.transaction.Transactional;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 
 /**
@@ -30,7 +35,7 @@ import java.util.List;
  */
 @RestController
 @RequestMapping("/api/partyManager")
-public class PartyManager implements cz.tacr.elza.api.controller.PartyManager<ParAbstractPartyVals> {
+public class PartyManager implements cz.tacr.elza.api.controller.PartyManager<ParAbstractParty> {
 
     @Autowired
     private AbstractPartyRepository abstractPartyRepository;
@@ -41,54 +46,74 @@ public class PartyManager implements cz.tacr.elza.api.controller.PartyManager<Pa
     @Autowired
     private PartyTypeRepository partyTypeRepository;
 
+
     @RequestMapping(value = "/getPartyTypes", method = RequestMethod.GET)
     @Override
-    public List<ParPartyType> getPartyTypes() {
-        return partyTypeRepository.findAll();
+    public List<ParPartyTypeExt> getPartyTypes() {
+        List<ParPartyType> all = partyTypeRepository.findAll();
+        List<ParPartySubtype> allSubtype = partySubtypeRepository.findAll();
+        Map<ParPartyType, List<ParPartySubtype>> map = new HashMap<>();
+
+        all.forEach((partyType) -> {
+            map.put(partyType, new ArrayList<>());
+        });
+
+        allSubtype.forEach((partySubtype) -> {
+            List<ParPartySubtype> list = map.get(partySubtype.getPartyType());
+            list.add(partySubtype);
+        });
+
+        List<ParPartyTypeExt> result = new ArrayList<>();
+        map.forEach((partyType, subtypes) -> {
+            ParPartyTypeExt partyTypeExt = new ParPartyTypeExt();
+            BeanUtils.copyProperties(partyType, partyTypeExt);
+            partyTypeExt.setPartySubTypeList(subtypes);
+            result.add(partyTypeExt);
+        });
+
+        return result;
+    }
+
+    @Transactional
+    private void updateAbstractParty(final ParAbstractParty source, final ParAbstractParty target) {
+        Assert.notNull(source.getPartySubtype(), "Není vyplněné partySubtype");
+        Assert.notNull(source.getRecord(), "Není vyplněné record");
+        Integer recordId = source.getRecord().getRecordId();
+        Integer partySubtypeId = source.getPartySubtype().getPartySubtypeId();
+        Assert.notNull(partySubtypeId, "Není vyplněné partySubtypeId");
+        Assert.notNull(recordId, "Není vyplněné recordId");
+
+        final ParPartySubtype partySubtype =
+                partySubtypeRepository.findOne(partySubtypeId);
+        final RegRecord record = recordRepository.findOne(recordId);
+        Assert.notNull(partySubtype,
+                "Nebyla nalezena ParPartySubtype s id " + partySubtypeId);
+        Assert.notNull(record, "Nebyla nalezena RegRecord s id " + recordId);
+
+        target.setPartySubtype(partySubtype);
+        target.setRecord(record);
     }
 
     @RequestMapping(value = "/insertAbstractParty", method = RequestMethod.PUT)
     @Transactional
     @Override
-    public ParAbstractParty insertAbstractParty(@RequestBody final ParAbstractPartyVals abstractParty) {
-        Assert.notNull(abstractParty.getPartySubtypeId(), "Není vyplněné partySubtypeId");
-        Assert.notNull(abstractParty.getRecordId(), "Není vyplněné recordId");
-
-        final ParPartySubtype partySubtype =
-                partySubtypeRepository.findOne(abstractParty.getPartySubtypeId());
-        final RegRecord record = recordRepository.findOne(abstractParty.getRecordId());
-        Assert.notNull(partySubtype,
-                "Nebyla nalezena ParPartySubtype s id " + abstractParty.getPartySubtypeId());
-        Assert.notNull(record, "Nebyla nalezena RegRecord s id " + abstractParty.getRecordId());
-
-        ParAbstractParty party = new ParAbstractParty();
-        party.setPartySubtype(partySubtype);
-        party.setRecord(record);
-        abstractPartyRepository.save(party);
-        return party;
+    public ParAbstractParty insertAbstractParty(@RequestBody final ParAbstractParty abstractParty) {
+        ParAbstractParty newParty = new ParAbstractParty();
+        updateAbstractParty(abstractParty, newParty);
+        abstractPartyRepository.save(newParty);
+        return newParty;
     }
 
     @RequestMapping(value = "/updateAbstractParty", method = RequestMethod.PUT)
     @Transactional
     @Override
-    public ParAbstractParty updateAbstractParty(
-            @RequestParam("abstractPartyId") final Integer abstractPartyId,
-            @RequestBody final ParAbstractPartyVals abstractParty) {
+    public ParAbstractParty updateAbstractParty(@RequestBody final ParAbstractParty abstractParty) {
+        Integer abstractPartyId = abstractParty.getAbstractPartyId();
         Assert.notNull(abstractPartyId);
-        Assert.notNull(abstractParty.getPartySubtypeId(), "Není vyplněné partySubtypeId");
-        Assert.notNull(abstractParty.getRecordId(), "Není vyplněné recordId");
-        final ParPartySubtype partySubtype =
-                partySubtypeRepository.findOne(abstractParty.getPartySubtypeId());
-        final RegRecord record = recordRepository.findOne(abstractParty.getRecordId());
-        Assert.notNull(partySubtype,
-                "Nebyla nalezena ParPartySubtype s id " + abstractParty.getPartySubtypeId());
-        Assert.notNull(record, "Nebyla nalezena RegRecord s id " + abstractParty.getRecordId());
-
         ParAbstractParty party = abstractPartyRepository.findOne(abstractPartyId);
         Assert.notNull(party, "Nebyla nalezena ParAbstractParty s id " + abstractPartyId);
-        party.setPartySubtype(partySubtype);
-        party.setRecord(record);
-        abstractPartyRepository.save(party);
+        updateAbstractParty(abstractParty, abstractParty);
+        abstractPartyRepository.save(abstractParty);
         return party;
     }
 
@@ -96,6 +121,10 @@ public class PartyManager implements cz.tacr.elza.api.controller.PartyManager<Pa
     @Transactional
     @Override
     public void deleteAbstractParty(@RequestParam("abstractPartyId") final Integer abstractPartyId) {
+        ParAbstractParty abstractParty = abstractPartyRepository.findOne(abstractPartyId);
+        if (abstractParty == null) {
+            return;
+        }
         Assert.notNull(abstractPartyId);
         abstractPartyRepository.delete(abstractPartyId);
     }
